tests/requirements-test.txt
requests
pytest
pytest-html
selenium
webdriver-manager
appium-python-client
